+++ tests/integration/tests/consistency.py
@@ -1,8 +1,8 @@
 import time
 from itertools import combinations
 from testflows.connect import Shell
-
-from integration.tests.steps import *
+from integration.tests.steps.sql import *
+from integration.tests.steps.service_settings_steps import *
 
 
 @TestOutline
