+++ tests/integration/tests/manual_section.py
@@ -1,7 +1,6 @@
 from itertools import combinations
-from testflows.connect import Shell
-
-from integration.tests.steps import *
+from integration.tests.steps.sql import *
+from integration.tests.steps.service_settings_steps import *
 
 
 @TestOutline
