+++ tests/integration/tests/truncate.py
@@ -1,4 +1,5 @@
-from integration.tests.steps import *
+from integration.tests.steps.sql import *
+from integration.tests.steps.service_settings_steps import *
 
 
 @TestScenario
